refactor: restController와 일반 controller를 분리한다
